+++ src/Service/S3/src/S3Client.php
@@ -3209,6 +3209,7 @@ class S3Client extends AbstractApi
             case 'ap-southeast-3':
             case 'ap-southeast-4':
             case 'ap-southeast-5':
+            case 'ap-southeast-7':
             case 'ca-central-1':
             case 'ca-west-1':
             case 'eu-central-1':